+++ src/ir_def/arrange_ir.ml
@@ -140,6 +140,7 @@ and pat p = match p.it with
   | OptP p          -> "OptP"       $$ [ pat p ]
   | TagP (i, p)     -> "TagP"       $$ [ id i; pat p ]
   | AltP (p1,p2)    -> "AltP"       $$ [ pat p1; pat p2 ]
+  | AndP (p1,p2)    -> "AndP"       $$ [ pat p1; pat p2 ]
 
 and lit = function
   | NullLit       -> Atom "NullLit"