+++ src/mo_def/arrange.ml
@@ -169,6 +169,7 @@ module Make (Cfg : Config) = struct
     | OptP p          -> "OptP"       $$ [pat p]
     | TagP (i, p)     -> "TagP"       $$ [tag i; pat p]
     | AltP (p1, p2)   -> "AltP"       $$ [pat p1; pat p2]
+    | AndP (p1, p2)   -> "AndP"       $$ [pat p1; pat p2]
     | ParP p          -> "ParP"       $$ [pat p]))
 
   and lit = function