+++ src/mo_frontend/definedness.ml
@@ -157,6 +157,7 @@ and pat msgs p : fd = match p.it with
   | OptP p
   | TagP (_, p)   -> pat msgs p
   | AltP (p1, p2) -> pat msgs p1 ++++ pat msgs p2
+  | AndP (p1, p2) -> pat msgs p1 ++++ pat msgs p2
 
 and pats msgs ps : fd = union_binders (pat msgs) ps