+++ src/mo_frontend/test_recovery.ml
@@ -293,6 +293,7 @@ let%expect_test "test3" =
       or <pat_bin> (e.g. 'or x')
       , seplist(<pat_bin>,,) (e.g. ', x')
       : <typ> (e.g. ': Int')
+      and <pat_bin> (e.g. 'and x')
 
     (unknown location): syntax error [M0001], unexpected token 'private', expected one of token or <phrase> sequence:
       }
@@ -300,7 +301,8 @@ let%expect_test "test3" =
 
     (unknown location): syntax error [M0001], unexpected token 'let', expected one of token or <phrase> sequence:
       }
-      ; seplist(<dec_field>,<semicolon>) (e.g. '; public let x : Int = 0') |}] 
+      ; seplist(<dec_field>,<semicolon>) (e.g. '; public let x : Int = 0')
+    |}] 
 
 
 let%expect_test "test4" =